backend-accounts/requirements.txt
aiofiles==23.2.1
dnspython==2.6.1
html5tagger==1.3.0
httptools==0.6.1
multidict==6.0.5
pika==1.3.2
pymongo==4.6.2
python-dotenv==1.0.1
sanic==23.12.1
sanic-routing==23.12.0
tracerite==1.1.1
typing_extensions==4.10.0
ujson==5.9.0
uvloop==0.19.0
websockets==12.0
